+++ Services/UsuarioService.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using ChronoLabel.Models;
 using ChronoLabel.Repositories;
+using DocumentValidator;
 
 namespace ChronoLabel.Services
 {
@@ -23,16 +24,70 @@ namespace ChronoLabel.Services
         }
         public void AddUsuario(Usuario usuario)
         {
+            if (usuario is null)
+            {
+                throw new ArgumentNullException(nameof(usuario), "Usuário não pode ser nulo.");
+            }
+
+            if(!CpfValidation.Validate(usuario.Cpf))
+            {
+                throw new ArgumentException("CPF inválido.", nameof(usuario.Cpf));
+            }
+
+            if (string.IsNullOrWhiteSpace(usuario.Cpf) ||
+            string.IsNullOrWhiteSpace(usuario.Nome) ||
+            string.IsNullOrWhiteSpace(usuario.Senha) ||
+            string.IsNullOrWhiteSpace(usuario.Tipo))
+            {
+                throw new ArgumentException("Todos os campos do usuário são obrigatórios.");
+            }
+
+            if (_usuarioRepository.UsuarioExists(usuario.Cpf))
+            {
+                throw new InvalidOperationException("Já existe um usuário com esse CPF.");
+            }
             _usuarioRepository.AddUsuario(usuario);
         }
 
         public void UpdateUsuario(Usuario usuario)
         {
-            _usuarioRepository.UpdateUsuario(usuario);
+            if (string.IsNullOrWhiteSpace(usuario.Cpf))
+            {
+                throw new ArgumentNullException(nameof(usuario), "CPF deve ser preenchido.");
+            }
+
+            var existingUsuario = _usuarioRepository.GetUsuarioByCpf(usuario.Cpf);
+
+            if (existingUsuario is null)
+            {
+                throw new InvalidOperationException("Usuário não encontrado.");
+            }
+
+            if (!string.IsNullOrWhiteSpace(usuario.Nome))
+            {
+                existingUsuario.Nome = usuario.Nome;
+            }
+
+            if (!string.IsNullOrWhiteSpace(usuario.Senha))
+            {
+                existingUsuario.Senha = usuario.Senha;
+            }
+
+            if (!string.IsNullOrWhiteSpace(usuario.Tipo))
+            {
+                existingUsuario.Tipo = usuario.Tipo;
+            }
+
+            _usuarioRepository.UpdateUsuario(existingUsuario);
         }
 
         public void DeleteUsuario(string cpf)
         {
+            if (!_usuarioRepository.UsuarioExists(cpf))
+            {
+                throw new InvalidOperationException("Usuário não encontrado.");
+            }
+
             _usuarioRepository.DeleteUsuario(cpf);
         }
